Changed: Parameter type Changed latitude and longitude type from Float to Double

+++ app/src/main/java/com/kennek/skyspy/data/sources/RemoteWeatherDataSource.kt
@@ -9,8 +9,8 @@ class RemoteWeatherDataSource @Inject constructor(
     private val api: WeatherApi
 ){
     suspend fun getForecast(
-        latitude: Float,
-        longitude: Float,
+        latitude: Double,
+        longitude: Double,
         current: String,
         hourly: String,
         daily: String,
